+++ package.json
@@ -14,10 +14,12 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@discordjs/opus": "^0.9.0",
     "@discordjs/voice": "^0.16.0",
     "discord.js": "^14.11.0",
     "dotenv": "^16.0.3",
     "ffmpeg-static": "^5.1.0",
+    "tweetnacl": "^1.0.3",
     "ytdl-core": "^4.11.4"
   },
   "devDependencies": {
