+++ src/commands/music/leave.ts
@@ -1,5 +1,6 @@
 import { ChatInputCommandInteraction, SlashCommandBuilder } from "discord.js";
 import { ChatCommandMetadata } from "../../types/CommandDTO.js";
+import { getVoiceConnection } from "@discordjs/voice";
 
 const data: ChatCommandMetadata = {
   builder: new SlashCommandBuilder()
@@ -7,10 +8,18 @@ const data: ChatCommandMetadata = {
     .setDescription("Shoos the Llama from any channels he is currently in."),
   action: async (interaction: ChatInputCommandInteraction) => {
     await interaction.deferReply();
+    if (!interaction.guild) {
+      await interaction.editReply(
+        "🛑This command can only be used in a guild."
+      );
+      return;
+    }
+
     const client = interaction.client;
-    if (client.voiceConnection) {
+    const connection = getVoiceConnection(interaction.guild?.id);
+    if (connection) {
       client.player?.pause();
-      client.voiceConnection.destroy();
+      connection.destroy();
       await interaction.editReply("The Llama has left the venue.🦙💨");
     } else {
       await interaction.editReply(
